+++ backend/pkg/messages/read-message.go
@@ -77,9 +77,9 @@ func DecodeSessionEndDeprecated(reader BytesReader) (Message, error) {
 	return msg, err
 }
 
-func DecodeSetPageLocation(reader BytesReader) (Message, error) {
+func DecodeSetPageLocationDeprecated(reader BytesReader) (Message, error) {
 	var err error = nil
-	msg := &SetPageLocation{}
+	msg := &SetPageLocationDeprecated{}
 	if msg.URL, err = reader.ReadString(); err != nil {
 		return nil, err
 	}
@@ -1032,6 +1032,30 @@ func DecodeMouseClick(reader BytesReader) (Message, error) {
 	if msg.Selector, err = reader.ReadString(); err != nil {
 		return nil, err
 	}
+	if msg.NormalizedX, err = reader.ReadUint(); err != nil {
+		return nil, err
+	}
+	if msg.NormalizedY, err = reader.ReadUint(); err != nil {
+		return nil, err
+	}
+	return msg, err
+}
+
+func DecodeMouseClickDeprecated(reader BytesReader) (Message, error) {
+	var err error = nil
+	msg := &MouseClickDeprecated{}
+	if msg.ID, err = reader.ReadUint(); err != nil {
+		return nil, err
+	}
+	if msg.HesitationTime, err = reader.ReadUint(); err != nil {
+		return nil, err
+	}
+	if msg.Label, err = reader.ReadString(); err != nil {
+		return nil, err
+	}
+	if msg.Selector, err = reader.ReadString(); err != nil {
+		return nil, err
+	}
 	return msg, err
 }
 
@@ -1423,6 +1447,24 @@ func DecodeRedux(reader BytesReader) (Message, error) {
 		return nil, err
 	}
 	if msg.ActionTime, err = reader.ReadUint(); err != nil {
+		return nil, err
+	}
+	return msg, err
+}
+
+func DecodeSetPageLocation(reader BytesReader) (Message, error) {
+	var err error = nil
+	msg := &SetPageLocation{}
+	if msg.URL, err = reader.ReadString(); err != nil {
+		return nil, err
+	}
+	if msg.Referrer, err = reader.ReadString(); err != nil {
+		return nil, err
+	}
+	if msg.NavigationStart, err = reader.ReadUint(); err != nil {
+		return nil, err
+	}
+	if msg.DocumentTitle, err = reader.ReadString(); err != nil {
 		return nil, err
 	}
 	return msg, err
@@ -1899,7 +1941,7 @@ func ReadMessage(t uint64, reader BytesReader) (Message, error) {
 	case 3:
 		return DecodeSessionEndDeprecated(reader)
 	case 4:
-		return DecodeSetPageLocation(reader)
+		return DecodeSetPageLocationDeprecated(reader)
 	case 5:
 		return DecodeSetViewportSize(reader)
 	case 6:
@@ -2012,8 +2054,10 @@ func ReadMessage(t uint64, reader BytesReader) (Message, error) {
 		return DecodeAssetCache(reader)
 	case 67:
 		return DecodeCSSInsertRuleURLBased(reader)
-	case 69:
+	case 68:
 		return DecodeMouseClick(reader)
+	case 69:
+		return DecodeMouseClickDeprecated(reader)
 	case 70:
 		return DecodeCreateIFrameDocument(reader)
 	case 71:
@@ -2064,6 +2108,8 @@ func ReadMessage(t uint64, reader BytesReader) (Message, error) {
 		return DecodeTagTrigger(reader)
 	case 121:
 		return DecodeRedux(reader)
+	case 122:
+		return DecodeSetPageLocation(reader)
 	case 125:
 		return DecodeIssueEvent(reader)
 	case 126: