+++ ee/connectors/msgcodec/messages.py
@@ -42,7 +42,7 @@ class SessionEndDeprecated(Message):
         self.timestamp = timestamp
 
 
-class SetPageLocation(Message):
+class SetPageLocationDeprecated(Message):
     __id__ = 4
 
     def __init__(self, url, referrer, navigation_start):
@@ -585,6 +585,18 @@ class CSSInsertRuleURLBased(Message):
 
 
 class MouseClick(Message):
+    __id__ = 68
+
+    def __init__(self, id, hesitation_time, label, selector, normalized_x, normalized_y):
+        self.id = id
+        self.hesitation_time = hesitation_time
+        self.label = label
+        self.selector = selector
+        self.normalized_x = normalized_x
+        self.normalized_y = normalized_y
+
+
+class MouseClickDeprecated(Message):
     __id__ = 69
 
     def __init__(self, id, hesitation_time, label, selector):
@@ -825,6 +837,16 @@ class Redux(Message):
         self.action_time = action_time
 
 
+class SetPageLocation(Message):
+    __id__ = 122
+
+    def __init__(self, url, referrer, navigation_start, document_title):
+        self.url = url
+        self.referrer = referrer
+        self.navigation_start = navigation_start
+        self.document_title = document_title
+
+
 class IssueEvent(Message):
     __id__ = 125
 
@@ -854,7 +876,7 @@ class SessionSearch(Message):
         self.partition = partition
 
 
-class IOSSessionStart(Message):
+class MobileSessionStart(Message):
     __id__ = 90
 
     def __init__(self, timestamp, project_id, tracker_version, rev_id, user_uuid, user_os, user_os_version, user_device, user_device_type, user_country):
@@ -870,14 +892,14 @@ class IOSSessionStart(Message):
         self.user_country = user_country
 
 
-class IOSSessionEnd(Message):
+class MobileSessionEnd(Message):
     __id__ = 91
 
     def __init__(self, timestamp):
         self.timestamp = timestamp
 
 
-class IOSMetadata(Message):
+class MobileMetadata(Message):
     __id__ = 92
 
     def __init__(self, timestamp, length, key, value):
@@ -887,7 +909,7 @@ class IOSMetadata(Message):
         self.value = value
 
 
-class IOSEvent(Message):
+class MobileEvent(Message):
     __id__ = 93
 
     def __init__(self, timestamp, length, name, payload):
@@ -897,7 +919,7 @@ class IOSEvent(Message):
         self.payload = payload
 
 
-class IOSUserID(Message):
+class MobileUserID(Message):
     __id__ = 94
 
     def __init__(self, timestamp, length, id):
@@ -906,7 +928,7 @@ class IOSUserID(Message):
         self.id = id
 
 
-class IOSUserAnonymousID(Message):
+class MobileUserAnonymousID(Message):
     __id__ = 95
 
     def __init__(self, timestamp, length, id):
@@ -915,7 +937,7 @@ class IOSUserAnonymousID(Message):
         self.id = id
 
 
-class IOSScreenChanges(Message):
+class MobileScreenChanges(Message):
     __id__ = 96
 
     def __init__(self, timestamp, length, x, y, width, height):
@@ -927,7 +949,7 @@ class IOSScreenChanges(Message):
         self.height = height
 
 
-class IOSCrash(Message):
+class MobileCrash(Message):
     __id__ = 97
 
     def __init__(self, timestamp, length, name, reason, stacktrace):
@@ -938,7 +960,7 @@ class IOSCrash(Message):
         self.stacktrace = stacktrace
 
 
-class IOSViewComponentEvent(Message):
+class MobileViewComponentEvent(Message):
     __id__ = 98
 
     def __init__(self, timestamp, length, screen_name, view_name, visible):
@@ -949,7 +971,7 @@ class IOSViewComponentEvent(Message):
         self.visible = visible
 
 
-class IOSClickEvent(Message):
+class MobileClickEvent(Message):
     __id__ = 100
 
     def __init__(self, timestamp, length, label, x, y):
@@ -960,7 +982,7 @@ class IOSClickEvent(Message):
         self.y = y
 
 
-class IOSInputEvent(Message):
+class MobileInputEvent(Message):
     __id__ = 101
 
     def __init__(self, timestamp, length, value, value_masked, label):
@@ -971,7 +993,7 @@ class IOSInputEvent(Message):
         self.label = label
 
 
-class IOSPerformanceEvent(Message):
+class MobilePerformanceEvent(Message):
     __id__ = 102
 
     def __init__(self, timestamp, length, name, value):
@@ -981,7 +1003,7 @@ class IOSPerformanceEvent(Message):
         self.value = value
 
 
-class IOSLog(Message):
+class MobileLog(Message):
     __id__ = 103
 
     def __init__(self, timestamp, length, severity, content):
@@ -991,7 +1013,7 @@ class IOSLog(Message):
         self.content = content
 
 
-class IOSInternalError(Message):
+class MobileInternalError(Message):
     __id__ = 104
 
     def __init__(self, timestamp, length, content):
@@ -1000,7 +1022,7 @@ class IOSInternalError(Message):
         self.content = content
 
 
-class IOSNetworkCall(Message):
+class MobileNetworkCall(Message):
     __id__ = 105
 
     def __init__(self, timestamp, length, type, method, url, request, response, status, duration):
@@ -1015,7 +1037,7 @@ class IOSNetworkCall(Message):
         self.duration = duration
 
 
-class IOSSwipeEvent(Message):
+class MobileSwipeEvent(Message):
     __id__ = 106
 
     def __init__(self, timestamp, length, label, x, y, direction):
@@ -1027,7 +1049,7 @@ class IOSSwipeEvent(Message):
         self.direction = direction
 
 
-class IOSBatchMeta(Message):
+class MobileBatchMeta(Message):
     __id__ = 107
 
     def __init__(self, timestamp, length, first_index):
@@ -1036,7 +1058,7 @@ class IOSBatchMeta(Message):
         self.first_index = first_index
 
 
-class IOSPerformanceAggregated(Message):
+class MobilePerformanceAggregated(Message):
     __id__ = 110
 
     def __init__(self, timestamp_start, timestamp_end, min_fps, avg_fps, max_fps, min_cpu, avg_cpu, max_cpu, min_memory, avg_memory, max_memory, min_battery, avg_battery, max_battery):
@@ -1056,7 +1078,7 @@ class IOSPerformanceAggregated(Message):
         self.max_battery = max_battery
 
 
-class IOSIssueEvent(Message):
+class MobileIssueEvent(Message):
     __id__ = 111
 
     def __init__(self, timestamp, type, context_string, context, payload):