+++ ee/connectors/msgcodec/messages.pyx
@@ -67,7 +67,7 @@ cdef class SessionEndDeprecated(PyMessage):
         self.timestamp = timestamp
 
 
-cdef class SetPageLocation(PyMessage):
+cdef class SetPageLocationDeprecated(PyMessage):
     cdef public int __id__
     cdef public str url
     cdef public str referrer
@@ -872,6 +872,25 @@ cdef class MouseClick(PyMessage):
     cdef public unsigned long hesitation_time
     cdef public str label
     cdef public str selector
+    cdef public unsigned long normalized_x
+    cdef public unsigned long normalized_y
+
+    def __init__(self, unsigned long id, unsigned long hesitation_time, str label, str selector, unsigned long normalized_x, unsigned long normalized_y):
+        self.__id__ = 68
+        self.id = id
+        self.hesitation_time = hesitation_time
+        self.label = label
+        self.selector = selector
+        self.normalized_x = normalized_x
+        self.normalized_y = normalized_y
+
+
+cdef class MouseClickDeprecated(PyMessage):
+    cdef public int __id__
+    cdef public unsigned long id
+    cdef public unsigned long hesitation_time
+    cdef public str label
+    cdef public str selector
 
     def __init__(self, unsigned long id, unsigned long hesitation_time, str label, str selector):
         self.__id__ = 69
@@ -1218,6 +1237,21 @@ cdef class Redux(PyMessage):
         self.action_time = action_time
 
 
+cdef class SetPageLocation(PyMessage):
+    cdef public int __id__
+    cdef public str url
+    cdef public str referrer
+    cdef public unsigned long navigation_start
+    cdef public str document_title
+
+    def __init__(self, str url, str referrer, unsigned long navigation_start, str document_title):
+        self.__id__ = 122
+        self.url = url
+        self.referrer = referrer
+        self.navigation_start = navigation_start
+        self.document_title = document_title
+
+
 cdef class IssueEvent(PyMessage):
     cdef public int __id__
     cdef public unsigned long message_id
@@ -1261,7 +1295,7 @@ cdef class SessionSearch(PyMessage):
         self.partition = partition
 
 
-cdef class IOSSessionStart(PyMessage):
+cdef class MobileSessionStart(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long project_id
@@ -1288,7 +1322,7 @@ cdef class IOSSessionStart(PyMessage):
         self.user_country = user_country
 
 
-cdef class IOSSessionEnd(PyMessage):
+cdef class MobileSessionEnd(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
 
@@ -1297,7 +1331,7 @@ cdef class IOSSessionEnd(PyMessage):
         self.timestamp = timestamp
 
 
-cdef class IOSMetadata(PyMessage):
+cdef class MobileMetadata(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1312,7 +1346,7 @@ cdef class IOSMetadata(PyMessage):
         self.value = value
 
 
-cdef class IOSEvent(PyMessage):
+cdef class MobileEvent(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1327,7 +1361,7 @@ cdef class IOSEvent(PyMessage):
         self.payload = payload
 
 
-cdef class IOSUserID(PyMessage):
+cdef class MobileUserID(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1340,7 +1374,7 @@ cdef class IOSUserID(PyMessage):
         self.id = id
 
 
-cdef class IOSUserAnonymousID(PyMessage):
+cdef class MobileUserAnonymousID(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1353,7 +1387,7 @@ cdef class IOSUserAnonymousID(PyMessage):
         self.id = id
 
 
-cdef class IOSScreenChanges(PyMessage):
+cdef class MobileScreenChanges(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1372,7 +1406,7 @@ cdef class IOSScreenChanges(PyMessage):
         self.height = height
 
 
-cdef class IOSCrash(PyMessage):
+cdef class MobileCrash(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1389,7 +1423,7 @@ cdef class IOSCrash(PyMessage):
         self.stacktrace = stacktrace
 
 
-cdef class IOSViewComponentEvent(PyMessage):
+cdef class MobileViewComponentEvent(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1406,7 +1440,7 @@ cdef class IOSViewComponentEvent(PyMessage):
         self.visible = visible
 
 
-cdef class IOSClickEvent(PyMessage):
+cdef class MobileClickEvent(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1423,7 +1457,7 @@ cdef class IOSClickEvent(PyMessage):
         self.y = y
 
 
-cdef class IOSInputEvent(PyMessage):
+cdef class MobileInputEvent(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1440,7 +1474,7 @@ cdef class IOSInputEvent(PyMessage):
         self.label = label
 
 
-cdef class IOSPerformanceEvent(PyMessage):
+cdef class MobilePerformanceEvent(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1455,7 +1489,7 @@ cdef class IOSPerformanceEvent(PyMessage):
         self.value = value
 
 
-cdef class IOSLog(PyMessage):
+cdef class MobileLog(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1470,7 +1504,7 @@ cdef class IOSLog(PyMessage):
         self.content = content
 
 
-cdef class IOSInternalError(PyMessage):
+cdef class MobileInternalError(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1483,7 +1517,7 @@ cdef class IOSInternalError(PyMessage):
         self.content = content
 
 
-cdef class IOSNetworkCall(PyMessage):
+cdef class MobileNetworkCall(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1508,7 +1542,7 @@ cdef class IOSNetworkCall(PyMessage):
         self.duration = duration
 
 
-cdef class IOSSwipeEvent(PyMessage):
+cdef class MobileSwipeEvent(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1527,7 +1561,7 @@ cdef class IOSSwipeEvent(PyMessage):
         self.direction = direction
 
 
-cdef class IOSBatchMeta(PyMessage):
+cdef class MobileBatchMeta(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public unsigned long length
@@ -1540,7 +1574,7 @@ cdef class IOSBatchMeta(PyMessage):
         self.first_index = first_index
 
 
-cdef class IOSPerformanceAggregated(PyMessage):
+cdef class MobilePerformanceAggregated(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp_start
     cdef public unsigned long timestamp_end
@@ -1575,7 +1609,7 @@ cdef class IOSPerformanceAggregated(PyMessage):
         self.max_battery = max_battery
 
 
-cdef class IOSIssueEvent(PyMessage):
+cdef class MobileIssueEvent(PyMessage):
     cdef public int __id__
     cdef public unsigned long timestamp
     cdef public str type