+++ ee/connectors/msgcodec/msgcodec.py
@@ -124,7 +124,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 4:
-            return SetPageLocation(
+            return SetPageLocationDeprecated(
                 url=self.read_string(reader),
                 referrer=self.read_string(reader),
                 navigation_start=self.read_uint(reader)
@@ -551,8 +551,18 @@ class MessageCodec(Codec):
                 base_url=self.read_string(reader)
             )
 
-        if message_id == 69:
+        if message_id == 68:
             return MouseClick(
+                id=self.read_uint(reader),
+                hesitation_time=self.read_uint(reader),
+                label=self.read_string(reader),
+                selector=self.read_string(reader),
+                normalized_x=self.read_uint(reader),
+                normalized_y=self.read_uint(reader)
+            )
+
+        if message_id == 69:
+            return MouseClickDeprecated(
                 id=self.read_uint(reader),
                 hesitation_time=self.read_uint(reader),
                 label=self.read_string(reader),
@@ -740,6 +750,14 @@ class MessageCodec(Codec):
                 action_time=self.read_uint(reader)
             )
 
+        if message_id == 122:
+            return SetPageLocation(
+                url=self.read_string(reader),
+                referrer=self.read_string(reader),
+                navigation_start=self.read_uint(reader),
+                document_title=self.read_string(reader)
+            )
+
         if message_id == 125:
             return IssueEvent(
                 message_id=self.read_uint(reader),
@@ -764,7 +782,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 90:
-            return IOSSessionStart(
+            return MobileSessionStart(
                 timestamp=self.read_uint(reader),
                 project_id=self.read_uint(reader),
                 tracker_version=self.read_string(reader),
@@ -778,12 +796,12 @@ class MessageCodec(Codec):
             )
 
         if message_id == 91:
-            return IOSSessionEnd(
+            return MobileSessionEnd(
                 timestamp=self.read_uint(reader)
             )
 
         if message_id == 92:
-            return IOSMetadata(
+            return MobileMetadata(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 key=self.read_string(reader),
@@ -791,7 +809,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 93:
-            return IOSEvent(
+            return MobileEvent(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 name=self.read_string(reader),
@@ -799,21 +817,21 @@ class MessageCodec(Codec):
             )
 
         if message_id == 94:
-            return IOSUserID(
+            return MobileUserID(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 id=self.read_string(reader)
             )
 
         if message_id == 95:
-            return IOSUserAnonymousID(
+            return MobileUserAnonymousID(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 id=self.read_string(reader)
             )
 
         if message_id == 96:
-            return IOSScreenChanges(
+            return MobileScreenChanges(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 x=self.read_uint(reader),
@@ -823,7 +841,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 97:
-            return IOSCrash(
+            return MobileCrash(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 name=self.read_string(reader),
@@ -832,7 +850,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 98:
-            return IOSViewComponentEvent(
+            return MobileViewComponentEvent(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 screen_name=self.read_string(reader),
@@ -841,7 +859,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 100:
-            return IOSClickEvent(
+            return MobileClickEvent(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 label=self.read_string(reader),
@@ -850,7 +868,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 101:
-            return IOSInputEvent(
+            return MobileInputEvent(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 value=self.read_string(reader),
@@ -859,7 +877,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 102:
-            return IOSPerformanceEvent(
+            return MobilePerformanceEvent(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 name=self.read_string(reader),
@@ -867,7 +885,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 103:
-            return IOSLog(
+            return MobileLog(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 severity=self.read_string(reader),
@@ -875,14 +893,14 @@ class MessageCodec(Codec):
             )
 
         if message_id == 104:
-            return IOSInternalError(
+            return MobileInternalError(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 content=self.read_string(reader)
             )
 
         if message_id == 105:
-            return IOSNetworkCall(
+            return MobileNetworkCall(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 type=self.read_string(reader),
@@ -895,7 +913,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 106:
-            return IOSSwipeEvent(
+            return MobileSwipeEvent(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 label=self.read_string(reader),
@@ -905,14 +923,14 @@ class MessageCodec(Codec):
             )
 
         if message_id == 107:
-            return IOSBatchMeta(
+            return MobileBatchMeta(
                 timestamp=self.read_uint(reader),
                 length=self.read_uint(reader),
                 first_index=self.read_uint(reader)
             )
 
         if message_id == 110:
-            return IOSPerformanceAggregated(
+            return MobilePerformanceAggregated(
                 timestamp_start=self.read_uint(reader),
                 timestamp_end=self.read_uint(reader),
                 min_fps=self.read_uint(reader),
@@ -930,7 +948,7 @@ class MessageCodec(Codec):
             )
 
         if message_id == 111:
-            return IOSIssueEvent(
+            return MobileIssueEvent(
                 timestamp=self.read_uint(reader),
                 type=self.read_string(reader),
                 context_string=self.read_string(reader),