+++ frontend/app/components/Session/Player/MobilePlayer/PlayerContent.tsx
@@ -10,7 +10,6 @@ import PlayerBlock from './PlayerBlock';
 
 const TABS = {
   EVENTS: 'Activity',
-  HEATMAPS: 'Click map',
 };
 
 interface IProps {