+++ tracker/tracker/src/common/messages.gen.ts
@@ -3,7 +3,7 @@
 
 export declare const enum Type {
   Timestamp = 0,
-  SetPageLocation = 4,
+  SetPageLocationDeprecated = 4,
   SetViewportSize = 5,
   SetViewportScroll = 6,
   CreateDocument = 7,
@@ -52,7 +52,8 @@ export declare const enum Type {
   TechnicalInfo = 63,
   CustomIssue = 64,
   CSSInsertRuleURLBased = 67,
-  MouseClick = 69,
+  MouseClick = 68,
+  MouseClickDeprecated = 69,
   CreateIFrameDocument = 70,
   AdoptedSSReplaceURLBased = 71,
   AdoptedSSInsertRuleURLBased = 73,
@@ -75,6 +76,7 @@ export declare const enum Type {
   CanvasNode = 119,
   TagTrigger = 120,
   Redux = 121,
+  SetPageLocation = 122,
 }
 
 
@@ -83,8 +85,8 @@ export type Timestamp = [
   /*timestamp:*/ number,
 ]
 
-export type SetPageLocation = [
-  /*type:*/ Type.SetPageLocation,
+export type SetPageLocationDeprecated = [
+  /*type:*/ Type.SetPageLocationDeprecated,
   /*url:*/ string,
   /*referrer:*/ string,
   /*navigationStart:*/ number,
@@ -432,6 +434,16 @@ export type MouseClick = [
   /*hesitationTime:*/ number,
   /*label:*/ string,
   /*selector:*/ string,
+  /*normalizedX:*/ number,
+  /*normalizedY:*/ number,
+]
+
+export type MouseClickDeprecated = [
+  /*type:*/ Type.MouseClickDeprecated,
+  /*id:*/ number,
+  /*hesitationTime:*/ number,
+  /*label:*/ string,
+  /*selector:*/ string,
 ]
 
 export type CreateIFrameDocument = [
@@ -595,6 +607,14 @@ export type Redux = [
   /*actionTime:*/ number,
 ]
 
+export type SetPageLocation = [
+  /*type:*/ Type.SetPageLocation,
+  /*url:*/ string,
+  /*referrer:*/ string,
+  /*navigationStart:*/ number,
+  /*documentTitle:*/ string,
+]
 
-type Message =  Timestamp | SetPageLocation | SetViewportSize | SetViewportScroll | CreateDocument | CreateElementNode | CreateTextNode | MoveNode | RemoveNode | SetNodeAttribute | RemoveNodeAttribute | SetNodeData | SetNodeScroll | SetInputTarget | SetInputValue | SetInputChecked | MouseMove | NetworkRequestDeprecated | ConsoleLog | PageLoadTiming | PageRenderTiming | CustomEvent | UserID | UserAnonymousID | Metadata | CSSInsertRule | CSSDeleteRule | Fetch | Profiler | OTable | StateAction | ReduxDeprecated | Vuex | MobX | NgRx | GraphQL | PerformanceTrack | StringDict | SetNodeAttributeDict | ResourceTimingDeprecated | ConnectionInformation | SetPageVisibility | LoadFontFace | SetNodeFocus | LongTask | SetNodeAttributeURLBased | SetCSSDataURLBased | TechnicalInfo | CustomIssue | CSSInsertRuleURLBased | MouseClick | CreateIFrameDocument | AdoptedSSReplaceURLBased | AdoptedSSInsertRuleURLBased | AdoptedSSDeleteRule | AdoptedSSAddOwner | AdoptedSSRemoveOwner | JSException | Zustand | BatchMetadata | PartitionedMessage | NetworkRequest | WSChannel | InputChange | SelectionChange | MouseThrashing | UnbindNodes | ResourceTiming | TabChange | TabData | CanvasNode | TagTrigger | Redux
+
+type Message =  Timestamp | SetPageLocationDeprecated | SetViewportSize | SetViewportScroll | CreateDocument | CreateElementNode | CreateTextNode | MoveNode | RemoveNode | SetNodeAttribute | RemoveNodeAttribute | SetNodeData | SetNodeScroll | SetInputTarget | SetInputValue | SetInputChecked | MouseMove | NetworkRequestDeprecated | ConsoleLog | PageLoadTiming | PageRenderTiming | CustomEvent | UserID | UserAnonymousID | Metadata | CSSInsertRule | CSSDeleteRule | Fetch | Profiler | OTable | StateAction | ReduxDeprecated | Vuex | MobX | NgRx | GraphQL | PerformanceTrack | StringDict | SetNodeAttributeDict | ResourceTimingDeprecated | ConnectionInformation | SetPageVisibility | LoadFontFace | SetNodeFocus | LongTask | SetNodeAttributeURLBased | SetCSSDataURLBased | TechnicalInfo | CustomIssue | CSSInsertRuleURLBased | MouseClick | MouseClickDeprecated | CreateIFrameDocument | AdoptedSSReplaceURLBased | AdoptedSSInsertRuleURLBased | AdoptedSSDeleteRule | AdoptedSSAddOwner | AdoptedSSRemoveOwner | JSException | Zustand | BatchMetadata | PartitionedMessage | NetworkRequest | WSChannel | InputChange | SelectionChange | MouseThrashing | UnbindNodes | ResourceTiming | TabChange | TabData | CanvasNode | TagTrigger | Redux | SetPageLocation
 export default Message