+++ tracker/tracker/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 14.0.0
+
+- titles for tabs
+- new `MouseClick` message to introduce heatmaps instead of clickmaps
+- crossdomain iframe tracking functionality
+
 # 13.0.2
 
 - more file extensions for canvas